Adicionando lógica de checkout para convidados, limpeza de carrinho após a compra e otimização na visualização do processOrder

+++ ecommerce/store/models.py
@@ -1,6 +1,7 @@
 from django.db import models
 from django.contrib.auth.models import User
 
+
 class Customer(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, null=True, blank=True)
     name = models.CharField(max_length=200, null=True)
@@ -12,7 +13,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price = models.DecimalField(max_digits=7, decimal_places=2)
+    price = models.FloatField()
     digital = models.BooleanField(default=False, null=True, blank=False)
     image = models.ImageField(null=True, blank=True)
 
@@ -38,7 +39,7 @@ class Order(models.Model):
 
     def __str__(self):
         return str(self.id)
-    
+
     @property
     def shipping(self):
         shipping = False
